Some housekeeping but mostly add a bunch of code to LocalCommunity._log_files() to write out simulated data in a useful format.

+++ MESS/Region.py
@@ -246,7 +246,7 @@ class Region(object):
                     self.paramsdict[param] = tup
 
             elif param == "sequence_length":
-                self.paramsdict[param] = float(newvalue)
+                self.paramsdict[param] = int(newvalue)
             else:
                 self.paramsdict[param] = newvalue
         except Exception as inst:
